New auth controller (keep old one too for now)
